Rename ShowControls.vue to TableControls.vue for better clarity
