+++ frontend/src/components/TableContols.vue
